Updated the OTel Metrics plugins to the org.opensearch.dataprepper package. (#1872)
